// apps/web/src/api/tipos.gerados.ts
export interface components {
    schemas: {
// ...
        CriarRegraPlantaoRequest: {
            /** Dia Semana */
            dia_semana: number;
            /**
             * Hora Inicio
             * Format: time
             */
            hora_inicio: string;
            /**
             * Hora Fim
             * Format: time
             */
            hora_fim: string;
            /** Valor Adicional */
            valor_adicional: number | string;
        };
// ...
        RegraPlantaoResponse: {
            /** Id */
            id: string;
            /** Dia Semana */
            dia_semana: number;
            /**
             * Hora Inicio
             * Format: time
             */
            hora_inicio: string;
            /**
             * Hora Fim
             * Format: time
             */
            hora_fim: string;
            /** Valor Adicional */
            valor_adicional: string;
            /** Ativo */
            ativo: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}